feat: replaced `DeviceEventEmitter` with `Context`
