+++ ts/components/Toast/types.ts
@@ -1,0 +1,13 @@
+import { HapticFeedbackTypes } from "react-native-haptic-feedback";
+import { IOIcons } from "../core/icons";
+
+export type ToastVariant = "neutral" | "error" | "info" | "success" | "warning";
+
+export type Toast = {
+  message: string;
+  variant?: ToastVariant;
+  icon?: IOIcons;
+  hapticFeedback?: keyof typeof HapticFeedbackTypes;
+};
+
+export type ToastOptions = Omit<Toast, "message">;
