+++ ts/components/core/typography/CTA.tsx
@@ -1,10 +1,8 @@
 import * as React from "react";
 import { IOFontFamily, IOFontWeight } from "../fonts";
 import { IOTheme, IOThemeLight } from "../variables/IOColors";
-import { useIOSelector } from "../../../store/hooks";
-import { isDesignSystemEnabledSelector } from "../../../store/reducers/persistedPreferences";
-import { ExternalTypographyProps, TypographyProps } from "./common";
 import { useTypographyFactory } from "./Factory";
+import { ExternalTypographyProps, TypographyProps } from "./common";
 
 type AllowedColors = IOTheme["textBody-default"] | "blueIO-850";
 type AllowedWeight = Extract<IOFontWeight, "SemiBold" | "Regular">;
@@ -17,9 +15,6 @@ type OwnProps = ExternalTypographyProps<
 export const ctaFontSize = 16;
 export const ctaLineHeight = 20;
 export const ctaDefaultColor: AllowedColors = IOThemeLight["textBody-default"];
-/* Legacy typographic styles */
-const ctaLegacyFontName: IOFontFamily = "TitilliumWeb";
-const ctaLegacyDefaultWeight: AllowedWeight = "Regular";
 /* New typographic styles */
 const ctaFontName: IOFontFamily = "ReadexPro";
 const ctaDefaultWeight: AllowedWeight = "Regular";
@@ -30,16 +25,11 @@ const ctaDefaultWeight: AllowedWeight = "Regular";
  * @param props
  * @constructor
  */
-export const CTA: React.FunctionComponent<OwnProps> = props => {
-  const isDesignSystemEnabled = useIOSelector(isDesignSystemEnabledSelector);
-
-  return useTypographyFactory<AllowedWeight, AllowedColors>({
+export const CTA: React.FunctionComponent<OwnProps> = props =>
+  useTypographyFactory<AllowedWeight, AllowedColors>({
     ...props,
-    defaultWeight: isDesignSystemEnabled
-      ? ctaDefaultWeight
-      : ctaLegacyDefaultWeight,
+    defaultWeight: ctaDefaultWeight,
     defaultColor: ctaDefaultColor,
-    font: isDesignSystemEnabled ? ctaFontName : ctaLegacyFontName,
+    font: ctaFontName,
     fontStyle: { fontSize: ctaFontSize, lineHeight: ctaLineHeight }
   });
-};
